refactor: 리졸버, Config 패키지 변경
#48
