+++ server/src/main/java/com/project/yozmcafe/controller/LoginArgumentResolver.java
@@ -1,4 +1,4 @@
-package com.project.yozmcafe;
+package com.project.yozmcafe.controller;
 
 import com.project.yozmcafe.domain.member.Member;
 import com.project.yozmcafe.domain.member.MemberRepository;
@@ -17,6 +17,7 @@ public class LoginArgumentResolver implements HandlerMethodArgumentResolver {
 
     private final JwtTokenProvider jwtTokenProvider;
     private final MemberRepository memberRepository;
+
     public LoginArgumentResolver(final JwtTokenProvider jwtTokenProvider, final MemberRepository memberRepository) {
         this.jwtTokenProvider = jwtTokenProvider;
         this.memberRepository = memberRepository;
@@ -31,10 +32,12 @@ public class LoginArgumentResolver implements HandlerMethodArgumentResolver {
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
         String header = webRequest.getHeader("Authorization");
 
-        if(header == null){
+        if (header == null) {
             throw new AuthenticationException("유효한 사용자만 접근 가능합니다.");
         }
         return memberRepository.findById(jwtTokenProvider.getPayload(header))
-                .orElseThrow(()-> new AuthenticationException("유효하지 않은 토큰입니다."));
+                .orElseThrow(() -> new AuthenticationException("유효하지 않은 토큰입니다."));
     }
+
+
 }
